+++ src/Symfony/Component/Cache/Tests/Adapter/PredisAdapterTest.php
@@ -44,6 +44,8 @@ class PredisAdapterTest extends AbstractRedisAdapterTest
             'persistent_id' => null,
             'read_timeout' => 0,
             'retry_interval' => 0,
+            'compression' => true,
+            'tcp_keepalive' => 0,
             'lazy' => false,
             'database' => '1',
             'password' => null,